merge series for the same path together. fixes #13
When there are multiple metricDefs for the same series name,
metric-tank will return a full set of datapoints padded with nulls for
each metricDef.  This commit merges these datapoints into a single
series. Each of the datapoints from Metric-Tank is scanned and the
first point that is not null is used.  If all points are null then
null is used.